requirements-with-deps.txt
autograd==1.5
matplotlib==3.6.2
numpy==1.23.5
pandas==1.5.2
scipy==1.9.3
seaborn==0.12.1
statsmodels==0.13.5
tensorflow-macos==2.10.0
tensorflow-probability==0.14.0